did some minor refactoring
