+++ data/commits.json
@@ -1,6 +1,14 @@
 [
   {
     "index": 1,
+    "commit": "commit feed up and working",
+    "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/6c3499f0ee915f9da598a601321472c1eee27c1f",
+    "author": "coelacanth7",
+    "author_url": "https://github.com/coelacanth7",
+    "sha": "6c3499f0ee915f9da598a601321472c1eee27c1f"
+  },
+  {
+    "index": 2,
     "commit": "got the data scrubbed",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/ed084727d0f56ff17d13e5826993d3df47ff252b",
     "author": "coelacanth7",
@@ -8,7 +16,7 @@
     "sha": "ed084727d0f56ff17d13e5826993d3df47ff252b"
   },
   {
-    "index": 2,
+    "index": 3,
     "commit": "starting to work on scrubbing data",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/4b80cbb4329b9a276f76a6b2c74429ca3ad36a5a",
     "author": "coelacanth7",
@@ -16,7 +24,7 @@
     "sha": "4b80cbb4329b9a276f76a6b2c74429ca3ad36a5a"
   },
   {
-    "index": 3,
+    "index": 4,
     "commit": "retrieving and console log data from api",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/f5be6f2d59aa91cdb16d27fde0847ecf2d33a7f3",
     "author": "coelacanth7",
@@ -24,7 +32,7 @@
     "sha": "f5be6f2d59aa91cdb16d27fde0847ecf2d33a7f3"
   },
   {
-    "index": 4,
+    "index": 5,
     "commit": "organized a few things",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/7d0c2f4366ad598020e5970dfbda26f3b5172a11",
     "author": "coelacanth7",
@@ -32,7 +40,7 @@
     "sha": "7d0c2f4366ad598020e5970dfbda26f3b5172a11"
   },
   {
-    "index": 5,
+    "index": 6,
     "commit": "i got some url data somewhat parsed",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/438cb2eec9daa900c5b922634c614a076fd0284e",
     "author": "coelacanth7",
@@ -40,7 +48,7 @@
     "sha": "438cb2eec9daa900c5b922634c614a076fd0284e"
   },
   {
-    "index": 6,
+    "index": 7,
     "commit": "basic server is functioning",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/1fbb7148117e2790cce8e6895f4e35889b031d37",
     "author": "coelacanth7",
@@ -48,7 +56,7 @@
     "sha": "1fbb7148117e2790cce8e6895f4e35889b031d37"
   },
   {
-    "index": 7,
+    "index": 8,
     "commit": "commit feed is up",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/cd9081f1c816ef3b875c250393c70d9c59b22a75",
     "author": "coelacanth7",
@@ -56,7 +64,7 @@
     "sha": "cd9081f1c816ef3b875c250393c70d9c59b22a75"
   },
   {
-    "index": 8,
+    "index": 9,
     "commit": "html bootstrap form set up",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/751e0f8b938be97e3afa9bac2467cf3ecee146fc",
     "author": "coelacanth7",
@@ -64,7 +72,7 @@
     "sha": "751e0f8b938be97e3afa9bac2467cf3ecee146fc"
   },
   {
-    "index": 9,
+    "index": 10,
     "commit": "set up a wrapper",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/fbdf0ea907f05aa30ab4cde912fa828f59fd41df",
     "author": "coelacanth7",
@@ -72,7 +80,7 @@
     "sha": "fbdf0ea907f05aa30ab4cde912fa828f59fd41df"
   },
   {
-    "index": 10,
+    "index": 11,
     "commit": "update readme",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/d152722bb3bad21696d242939c9d46b07c40a85f",
     "author": "coelacanth7",
@@ -80,7 +88,7 @@
     "sha": "d152722bb3bad21696d242939c9d46b07c40a85f"
   },
   {
-    "index": 11,
+    "index": 12,
     "commit": "set up file structure",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/a6426b215c5d05f0c41934f9ecf020d592d6b4a5",
     "author": "coelacanth7",
@@ -88,7 +96,7 @@
     "sha": "a6426b215c5d05f0c41934f9ecf020d592d6b4a5"
   },
   {
-    "index": 12,
+    "index": 13,
     "commit": "Update README.md",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/ceec1b2306888905ae12750558706174132a6741",
     "author": "BideoWego",
@@ -96,7 +104,7 @@
     "sha": "ceec1b2306888905ae12750558706174132a6741"
   },
   {
-    "index": 13,
+    "index": 14,
     "commit": "Initial commit",
     "html_url": "https://github.com/coelacanth7/project_github_commit_feed/commit/526145a8b50b13bafc85f9c8e3fa603c93912f55",
     "author": "BideoWego",
